ACE_SSL enabled, debugged (non-secure socket optional)
git-svn-id: https://vast.svn.sourceforge.net/svnroot/vast@190 e31416da-b748-0410-9b4e-cac2f8189b79

+++ trunk/VAST/Demo/VASTsim_gui/VASTsim_gui.cpp
@@ -216,7 +216,7 @@ VOID Render (HWND hWnd)
         {
             // change color ?
             SelectObject (hdc, GetStockObject(HOLLOW_BRUSH));
-            Ellipse (hdc, x-aoi, y-aoi, x+aoi, y+aoi);
+            Ellipse (hdc, (int)(x-aoi), (int)(y-aoi), (int)(x+aoi), (int)(y+aoi));
         }
 
         if (i == g_selected)
@@ -233,7 +233,7 @@ VOID Render (HWND hWnd)
             if (show_self)
             {
                 //SelectObject( hdc, GetStockObject(HOLLOW_BRUSH) );
-                Ellipse (hdc, x-aoi, y-aoi, x+aoi, y+aoi);
+                Ellipse (hdc, (int)(x-aoi), (int)(y-aoi), (int)(x+aoi), (int)(y+aoi));
             }
 
             // draw neighbor dots
@@ -339,7 +339,7 @@ VOID Render (HWND hWnd)
              g_nodes_active, 
              selected_x, 
              selected_y, 
-             g_aoi, 
+             (int)g_aoi, 
              size_AN, 
              str2, 
              (follow_mode ? "[follow]" : ""),
