+++ trunk/VAST/Demo/demo_chatva/chatva.cpp
@@ -146,7 +146,7 @@ VOID Render (HWND hWnd)
 
     // draw AOI
     //SelectObject( hdc, GetStockObject(HOLLOW_BRUSH) );
-    Ellipse (hdc, (int)g_aoi.center.x - g_aoi.radius, (int)g_aoi.center.y - g_aoi.radius, (int)g_aoi.center.x + g_aoi.radius, (int)g_aoi.center.y + g_aoi.radius);
+    Ellipse (hdc, (int)(g_aoi.center.x - g_aoi.radius), (int)(g_aoi.center.y - g_aoi.radius), (int)(g_aoi.center.x + g_aoi.radius), (int)(g_aoi.center.y + g_aoi.radius));
     //Ellipse( hdc, x-aoi_buf, y-aoi_buf, x+aoi_buf, y+aoi_buf );        
 
     // draw matcher AOI, if available
@@ -154,7 +154,7 @@ VOID Render (HWND hWnd)
     if (matcher_aoi != NULL)
     {
         hPenOld = (HPEN)SelectObject (hdc, hPenRed);
-        Ellipse (hdc, (int)matcher_aoi->center.x - matcher_aoi->radius, (int)matcher_aoi->center.y - matcher_aoi->radius, (int)matcher_aoi->center.x + matcher_aoi->radius, (int)matcher_aoi->center.y + matcher_aoi->radius);
+        Ellipse (hdc, (int)(matcher_aoi->center.x - matcher_aoi->radius), (int)(matcher_aoi->center.y - matcher_aoi->radius), (int)(matcher_aoi->center.x + matcher_aoi->radius), (int)(matcher_aoi->center.y + matcher_aoi->radius));
         SelectObject (hdc, hPenOld);
     }
 
@@ -212,7 +212,7 @@ VOID Render (HWND hWnd)
              g_cursor.x-g_origin.x, g_cursor.y-g_origin.y, 
              (int)VASTnet::resolvePort (self->id)-g_netpara.port+1, (int)g_aoi.center.x, (int)g_aoi.center.y, 
              join_state,
-             (matcher_aoi == NULL ? self->aoi.radius : matcher_aoi->radius), 
+             (matcher_aoi == NULL ? (int)self->aoi.radius : (int)matcher_aoi->radius), 
              n, (follow_mode ? "[follow]" : ""), 
              last_char,
              GWstr.c_str ());
